# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SafeWalk API
  version: 1.0.0
  description: API REST para el sistema SafeWalk U de rutas seguras para estudiantes.

servers:
  - url: http://localhost:3000/api
    description: Servidor local

tags:
  - name: Auth
    description: Autenticación de usuarios
  - name: Usuarios
    description: Gestión de usuarios
  - name: Reportes
    description: Gestión de reportes de incidentes
  - name: Rutas
    description: Gestión de rutas seguras
  - name: Evidencias
    description: Gestión de evidencias multimedia

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required:
        - nombre
        - apellido
        - correo
        - contrasena
        - rol
      properties:
        nombre:
          type: string
          example: Milena
        apellido:
          type: string
          example: Ordoñez
        correo:
          type: string
          format: email
          example: milena@gmail.com
        contrasena:
          type: string
          example: "123456"
        rol:
          type: string
          enum: [ESTUDIANTE, ADMINISTRADOR]
          example: ADMINISTRADOR

    LoginRequest:
      type: object
      required:
        - correo
        - contrasena
      properties:
        correo:
          type: string
          format: email
          example: milena@gmail.com
        contrasena:
          type: string
          example: "123456"

    Usuario:
      type: object
      properties:
        id_usuario:
          type: integer
          example: 1
        nombre:
          type: string
          example: Milena
        apellido:
          type: string
          example: Ordoñez
        correo:
          type: string
          example: milena@gmail.com
        rol:
          type: string
          enum: [ESTUDIANTE, ADMINISTRADOR]
        estado:
          type: string
          enum: [ACTIVO, INACTIVO]

    ReporteRequest:
      type: object
      required:
        - descripcion
        - nivel_riesgo
        - id_usuario
        - id_ubicacion
      properties:
        descripcion:
          type: string
          example: Zona con poca iluminación cerca de la universidad
        nivel_riesgo:
          type: string
          enum: [BAJO, MEDIO, ALTO]
          example: ALTO
        id_usuario:
          type: integer
          example: 1
        id_ubicacion:
          type: integer
          example: 1

    ReporteUpdate:
      type: object
      required:
        - descripcion
        - nivel_riesgo
        - estado
      properties:
        descripcion:
          type: string
          example: Reporte actualizado
        nivel_riesgo:
          type: string
          enum: [BAJO, MEDIO, ALTO]
          example: MEDIO
        estado:
          type: string
          enum: [PENDIENTE, VALIDADO, RECHAZADO, DUPLICADO]
          example: VALIDADO

    RutaRequest:
      type: object
      required:
        - nombre_ruta
        - nivel_seguridad
        - tiempo_estimado
        - ubicaciones
        - puntos
      properties:
        nombre_ruta:
          type: string
          example: Ruta UTPL
        descripcion:
          type: string
          example: Ruta segura hacia la universidad
        nivel_seguridad:
          type: string
          enum: [BAJO, MEDIO, ALTO]
          example: ALTO
        tiempo_estimado:
          type: number
          example: 15
        ubicaciones:
          type: array
          minItems: 2
          maxItems: 50
          items:
            type: integer
          example: [1, 7]
        puntos:
          type: array
          minItems: 2
          maxItems: 500
          items:
            $ref: "#/components/schemas/RutaPunto"

    RutaPunto:
      type: object
      required:
        - latitud
        - longitud
      properties:
        latitud:
          type: number
          minimum: -90
          maximum: 90
        longitud:
          type: number
          minimum: -180
          maximum: 180
        tipo:
          type: string
          enum: [INICIO, INTERMEDIO, CRUCE, APOYO, DESTINO]
        observacion:
          type: string
          maxLength: 255

    EvidenciaRequest:
      type: object
      required:
        - url_archivo
        - tipo_archivo
        - id_reporte
      properties:
        url_archivo:
          type: string
          example: https://imagenes.com/foto.jpg
        tipo_archivo:
          type: string
          enum: [IMAGEN, VIDEO]
          example: IMAGEN
        id_reporte:
          type: integer
          example: 1

    Error401:
      type: object
      properties:
        message:
          type: string
          example: Token no enviado.

    Error403:
      type: object
      properties:
        message:
          type: string
          example: No tiene permisos.

    Error422:
      type: object
      properties:
        success:
          type: boolean
          example: false
        errors:
          type: object

paths:
  /health:
    get:
      tags:
        - Auth
      summary: Verificar estado de la API
      responses:
        "200":
          description: API funcionando correctamente

  /auth/register:
    post:
      tags:
        - Auth
      summary: Registrar un usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Usuario registrado correctamente
        "422":
          description: Error de validación Zod
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error422"

  /auth/login:
    post:
      tags:
        - Auth
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login exitoso
        "401":
          description: Credenciales inválidas
        "422":
          description: Error de validación Zod

  /users:
    get:
      tags:
        - Usuarios
      summary: Obtener todos los usuarios
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de usuarios
        "401":
          description: Token no enviado o inválido
        "403":
          description: No autorizado

  /users/{id}:
    get:
      tags:
        - Usuarios
      summary: Obtener usuario por ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Usuario encontrado
        "404":
          description: Usuario no encontrado

    put:
      tags:
        - Usuarios
      summary: Actualizar usuario
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Usuario"
      responses:
        "200":
          description: Usuario actualizado correctamente
        "401":
          description: Token no enviado o inválido
        "403":
          description: No autorizado
        "422":
          description: Error de validación Zod

    delete:
      tags:
        - Usuarios
      summary: Desactivar usuario mediante borrado lógico
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Usuario desactivado correctamente
        "401":
          description: Token no enviado o inválido
        "403":
          description: No autorizado

  /reports:
    get:
      tags:
        - Reportes
      summary: Obtener todos los reportes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de reportes

    post:
      tags:
        - Reportes
      summary: Crear reporte
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReporteRequest"
      responses:
        "201":
          description: Reporte creado correctamente
        "401":
          description: Token no enviado o inválido
        "403":
          description: No autorizado
        "422":
          description: Error de validación Zod

  /reports/{id}:
    get:
      tags:
        - Reportes
      summary: Obtener reporte por ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reporte encontrado
        "404":
          description: Reporte no encontrado

    put:
      tags:
        - Reportes
      summary: Actualizar reporte
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReporteUpdate"
      responses:
        "200":
          description: Reporte actualizado correctamente
        "403":
          description: No autorizado
        "422":
          description: Error de validación Zod

    delete:
      tags:
        - Reportes
      summary: Desactivar reporte mediante borrado lógico
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reporte desactivado correctamente
        "403":
          description: No autorizado

  /routes:
    get:
      tags:
        - Rutas
      summary: Obtener todas las rutas seguras
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de rutas

    post:
      tags:
        - Rutas
      summary: Crear ruta segura
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RutaRequest"
      responses:
        "201":
          description: Ruta creada correctamente
        "403":
          description: No autorizado
        "422":
          description: Error de validación Zod

  /routes/{id}:
    get:
      tags:
        - Rutas
      summary: Obtener ruta por ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ruta encontrada
        "404":
          description: Ruta no encontrada

    put:
      tags:
        - Rutas
      summary: Actualizar ruta segura
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RutaRequest"
      responses:
        "200":
          description: Ruta actualizada correctamente
        "403":
          description: No autorizado
        "422":
          description: Error de validación Zod

    delete:
      tags:
        - Rutas
      summary: Eliminar ruta segura
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ruta eliminada correctamente
        "403":
          description: No autorizado

  /evidencias:
    get:
      tags:
        - Evidencias
      summary: Obtener todas las evidencias
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de evidencias
        "403":
          description: No autorizado

    post:
      tags:
        - Evidencias
      summary: Crear evidencia
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenciaRequest"
      responses:
        "201":
          description: Evidencia creada correctamente
        "422":
          description: Error de validación Zod

  /evidencias/{id}:
    get:
      tags:
        - Evidencias
      summary: Obtener evidencia por ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Evidencia encontrada
        "404":
          description: Evidencia no encontrada

    put:
      tags:
        - Evidencias
      summary: Actualizar evidencia
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenciaRequest"
      responses:
        "200":
          description: Evidencia actualizada correctamente
        "403":
          description: No autorizado
        "422":
          description: Error de validación Zod

    delete:
      tags:
        - Evidencias
      summary: Eliminar evidencia
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Evidencia eliminada correctamente
        "403":
          description: No autorizado